Bump rusqlite to 0.39 and add r2d2 connection pooling deps
Update rusqlite 0.37 -> 0.39 to match r2d2_sqlite 0.34, and add r2d2 +
r2d2_sqlite as optional deps under the sqlite feature. rusqlite 0.39 dropped
ToSql for usize (platform-ambiguous width), so the migrator binds the
migration version as i64.

+++ src/database/adapters/sqlite/migrator.rs
@@ -226,7 +226,7 @@ impl<'a> Migrator<'a> {
                     INSERT INTO migrations (version, name, migrated_at) \
                     VALUES ($1, $2, $3)\
                 ",
-            params![migration.version, migration.name, Utc::now()],
+            params![migration.version as i64, migration.name, Utc::now()],
         )?;
 
         Ok(())
@@ -242,7 +242,7 @@ impl<'a> Migrator<'a> {
                     DELETE FROM migrations \
                     WHERE version = $1\
                 ",
-            params![migration.version],
+            params![migration.version as i64],
         )?;
 
         Ok(())
